+++ core/model/src/commonMain/kotlin/org/mifos/mobile/core/model/entity/templates/savings/ChargeOptions.kt
@@ -20,7 +20,7 @@ data class ChargeOptions(
     val active: Boolean? = null,
     val penalty: Boolean? = null,
     val currency: Currency? = null,
-    val amount: Float = 0.toFloat(),
+    val amount: Float = 0f,
     val chargeTimeType: ChargeTimeType,
     val chargeAppliesTo: ChargeAppliesTo,
     val chargeCalculationType: ChargeCalculationType,
